♻ `CaseHelper.copy_cases()` to `api._copy_cases()`
